fs: rework sprite loading to avoid redundancy
sce.db now has obj records specifying number of rotations to load,
and has spr records which list images as frames of an animation.
these images can thus be reused.
this does not deal with rotation images, which still contain
duplicates.
unfortunately, we can't reduce number of rotations on units
despite redundancies since the units we have have 17, not 16
unique images.
objects are assumed to have the same number of rotations for all
sprite sets.

like terrain, sprites are loaded on demand; both are now allocated
and referenced in linked lists during loading, and the pics loaded
afterwards in initimg, without modifying references in the main
data structures.
pics are allocated as arrays of pointers to Pic structs, which are
later filled in by loadpic.
this is still a bit fragile as there are assumptions in the loading
code about the size of the arrays, depending on type.