sim: use correct grid for movement and adjust deceleration treshold
using the pathfinding grid for movement means that we'll stop
moving as soon as target path coordinates are reached, but
that means we'd undershoot the target if we only touch the
path node boundary.
use pixel grid instead.

adjustment of deceleration heuristic needed to prevent stopping
before target is reached.