grp: experimental pcx palette support
unless we do palette remapping like scbw does, which would
add a ton of complexity to the drawing code, as well as
essentially requiring a special image format or using the
grp files directly and modifying them at runtime, we can
only guess how to use partially transparent sprites.

after countless experiments, i've settled on just generating
a best-effort image with an alpha channel, for which we'll
just do alpha-blending in drw, something simple and easy to
implement.
palette remapping is undocumented, but i'm guessing that the
basic palette is on the first row, and the sprite indices are
actually palette indices, changing the pixel value to a
precomputed value taking into account the  transparency  in
the sprite.  these palettes are per-tileset, another source
of complexity.
with trial and error, we select the column with the best
looking color, then compute transparency values by fitting an
exponential curve with a plateau, to emulate a bias towards
highly opaque values on the upper end and highly transparent
values on the low end.  it looks sort of ok.

overall, the seems like the simplest approach to the problem,
but maybe there exists a better one.
i've failed to identify a better trade-off between simplicity
and graphical fidelity.