fs: remove gfx db records, useless
the idea was to spawn a mutable object on top of the moving
unit, but it's really pointless and more complicated than
just an additional sprite layer in the draw order.
the sprite we're trying to add is scv's engine glow, which
is an effect other units may have in other ways, so call it
that.
generalize loading and drawing the various layers of the
same pic, record whether sprite has team colors or not.
when fixing obj sprites, skip buildings and set shadows as
well.
rename tscglow to an scv glow sprite, this simplifies the
db records.