Tmove/Tmovenear and protocol adjustments
- Tmove to move to target node and Tmovenear to move towards target
object; these two are different in that clicking inside an object
implies finding the closest free and reachable spot near the object,
which jps cannot handle.  we already employ some shitty heuristics
which don't really work well.  the pathfinding system will have to
be changed to handle both cases since one is a general case of the
other.
- terminate messages
- put message size inside header, check it, and use that; we used
to attempt to parse the entire message buffer, which is obviously
wrong
- T* values should not begin with 0 to avoid weird undetected bugs
- allow spaces in packing strings for readability