utils: sprite offsets may be 0, and crop(1) doesn't like that
the real fix might be to modify crop(1) instead.  it checks if
anything was actually specified at startup and compares
translate offsets to ZP.  if true it assumes usage error.

technically, it makes sense to treat noops as a usage error,
but you lose writing the same image to stdout.  given a file
argument, it's fine, but breaks pipelines.  perhaps it should
just ignore noop's, at least if there is no file argument.