begin rewrite of extraction scripts: add genspr2, scespr, scesha
- genspr2 will replace genspr as the main script; it does setup
  and performs the whole extraction
- scespr: generate a team color sprite sheet for a frame
- scesha: generate a shadow sprite from a frame image

deprecated: sceass, scerot

this implements the last insights on offset calculations.
grp files have a canvas width and height within which each frame
is painted, which was dismissed before.  in fact, these dimensions
must be known to calculate offsets for mirror rotations.

previously, when we generated rotations, we just set the same
offsets as the mirrored frame, which almost always results in an
incorrect x offset.  as vspr shows, only generated rotations have
wrong offsets.  the x offset correction is simply the slack space
between the right edge of the original sprite and the canvas width:

	Δx = canw - r.min.x - r.max.x

this takes care of mirror frames offsets.  next, there's a global
offset for drawing the frame.  this information is not in the grp
file but in unit tables.  units have w·h size in pixels and D/L/R/U
offsets from center in pixels.  offset for units looks to be the
difference between half the canvas width and the L offset, for both
dimensions:

	gΔx = gΔy = canw / 2 - L

for buildings, it seems to be multiples of -32: if global offset for
a dimensions is ≥ 32, it is set to -32:

	gΔx = canw / 2 - L >= 32 ? -32 : 0
	gΔy = canh / 2 - U >= 32 ? -32 : 0

shadows may have x and y offsets.  objects with a shadow sprite have
arbitrary offsets set somewhere else (dat files etc.), those have to
be either found prior to this or eyeballed.  the others have either
implied offsets or ones that are set elsewhere.  so far for ground
units it's (0,7) and (0,42) for air.  the buildings have different
ones that have been eyeballed with vspr.

this patch implements the basis of this all with a rewrite of genspr
and helper scripts.
remaining:
- sprites with transparency
- buildings, separate shadow sprites
- tileset
- clean(er) namespace manipulation
- testing

*hopefully*, it will all work out and we'll be able to fucking go on
to more interesting stuff!