refactoring: use actual vectors and Points
- a bunch of code duplicated what are essentially resizable
arrays; i don't like casting pointers, but not sure what's a
better way
- for pathing, just traverse vector backwards instead of
pushing points in a complicated way
- using Points right now simplifies and clarifies a lot of
code, there's no reason not to use them
- com: stop parsing message if one was invalid since we won't
even get an offset
- don't use anonymous Path struct, names are too confusing

all of this makes a bunch of code easier to look at and
arguably less bug-prone.

regarding the "hashing": an actual hash table wouldn't really
provide any benefits or more guarantees.
the Mobj references (idx/uuid) are a poor man's hash table,
but it's a simple solution to this.